src/dev_requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
Deprecated==1.2.10
Django==3.1.3
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-environ==0.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
et-xmlfile==1.1.0
flake8==3.8.4
future==0.18.2
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
pathspec==0.8.1
pep8==1.7.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGithub==1.53
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.10
pytz==2020.4
regex==2020.11.13
reportlab==3.5.55
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
